Basic Gravity implemented
Gravity is being applied to all models that dont have the boolean of
"isAffectedByGravity" set to false. Weight is a variable but needs to be
fine tuned and will be implemented when x and z accelerations come into
play. Perhaps a "Motion handler" that will call the gravity handler as
well as all of the other sub motion handlers.